Calculate growing season monthly means

Gowing seasons are approximately offset by 12 months between the Northern and Southern Hemispheres. Although this is a gross simplification, when trying to compare monthly values globally, it is necessary to align the summers and the winters, reffered to in this context as the growing seasons. This function aims to achive this by shifting the Southern hemishere values of a given grid by 6 months.

transform_monthly_means_into_growing_season_means(brick_of_monthly_means)

Arguments

brick_of_monthly_means
output from brick_calc_monthly_mean. A raster brick of 12 layers representing the 12 months of the year.

Value

a raster brick object with 12 layers, indicating the 12 months of the year, starting in the middle of winter.

Examples

#I need to add an exmple...